requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
pandas==2.1.3
numpy==1.25.2
yfinance==0.2.28
backtrader==1.9.78.123
ta-lib==0.4.28
pandas_ta==0.3.14b0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
python-multipart==0.0.6
APScheduler==3.10.4
structlog==23.2.0